Made location.js and moved code from currency to fix issues
